requirements.txt
pyautogen==0.2.25
chess==1.10.0
matplotlib
numpy
pandas
yfinance
python-dotenv

